+++ NEWS.md
@@ -1,3 +1,12 @@
+# finalfit 0.8.9
+
+* `ff_plot()` added
+* `coefficient_plot()` added
+* `variable_type()` added
+* Compatibility for future `shinyfit` started. 
+* `ff_relabel()` added.
+* Error added to `finalfit()` for not-allowed colons (:) in factor levels. #10
+
 # finalfit 0.8.8
 
 * `ff_glimpse()` re-written to remove `psych` dependency